+++ esperanto-tools-libs/CHANGELOG.md
@@ -16,7 +16,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.10.0]
 ### Added
 - [SW-15675] Added easyprofiler support.
-- [SW-16963] implements new requests for the server-client protocol; also improves the protocol versioning system.
+- [SW-16963] new requests for the server-client protocol; also improves the protocol versioning system.
+- [SW-17011] new IMonitor interface.
+- [SW-17066] unit tests for the new IMonitor interface.
 ## Fixed
  - [SW-16519] Fixed an issue freeing events in the multiprocess flow which made each subsequent waitForStream (for any stream) take longer and longer.